优化支付和订单服务

+++ youlexuan-pojo/src/main/java/com/zl/pojo/TbPayLog.java
@@ -1,16 +1,19 @@
 package com.zl.pojo;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 import java.util.Date;
 
 public class TbPayLog implements Serializable {
+
+    //订单编号
     private String outTradeNo;
 
     private Date createTime;
 
     private Date payTime;
 
-    private Long totalFee;
+    private BigDecimal totalFee;
 
     private String userId;
 
@@ -48,11 +51,12 @@ public class TbPayLog implements Serializable {
         this.payTime = payTime;
     }
 
-    public Long getTotalFee() {
+
+    public BigDecimal getTotalFee() {
         return totalFee;
     }
 
-    public void setTotalFee(Long totalFee) {
+    public void setTotalFee(BigDecimal totalFee) {
         this.totalFee = totalFee;
     }
